requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
minio==7.2.5
minio==7.2.7
